Removed interface PatchingHarvestLogManager
